requirements.txt
crewai
crewai_tools
langchain_community
langchain-together
langchain-google-genai
streamlit
